chore(components): update checkbox imports
